+++ libcef/browser/download_resume_util.h
@@ -14,7 +14,7 @@ namespace download_resume_util {
 // resume an interrupted download.
 void ResumeDownloadWithId(
     content::DownloadManager *manager, const std::string &guid, const GURL &url,
-    const base::FilePath &full_path, int received_bytes, int total_bytes,
+    const base::FilePath &full_path, int64_t received_bytes, int64_t total_bytes,
     const std::string &etag, const std::string &mime_type,
     const std::string &last_modified,
     std::vector<download::DownloadItem::ReceivedSlice> received_slices,
